+++ spectral_connectivity/__init__.py
@@ -5,7 +5,7 @@ directed connectivity measures from time series data using multitaper methods.
 """
 
 # flake8: noqa
-from spectral_connectivity.connectivity import Connectivity
+from spectral_connectivity.connectivity import Connectivity, MultivariateConnectivityResult
 from spectral_connectivity.minimum_phase_decomposition import (
     minimum_phase_reconstruction_error,
 )
@@ -47,6 +47,7 @@ __all__ = [
     "DEFAULT_METHODS",
     "JackknifeResult",
     "MorletWavelet",
+    "MultivariateConnectivityResult",
     "Multitaper",
     "MultitaperParameters",
     "ShortTimeFourierTransform",